fix(integrations): add scoped Kimi MCP attachment
